backend/requirements.txt
flask
flask-restful
flask-restful-swagger
flask-sqlalchemy
sqlalchemy-utils
passlib
requests
uwsgi
python-jose
pytest
psycopg2